requirements.txt
tqdm
numpy
scipy
numba
smplx
gdown
joblib
pytube
trimesh
pyrender
progress
filterpy
matplotlib
torchvision
scikit-image
scikit-video
pillow==6.2.1
opencv-python
torchsummary
ffmpeg
pyglet
git+https://github.com/mattloper/chumpy.git
git+https://github.com/mkocabas/yolov3-pytorch.git
git+https://github.com/mkocabas/multi-person-tracker.git
